// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ExpenseType = {
  fixed: 'fixed',
  variable: 'variable',
  installment: 'installment'
} as const

export type ExpenseType = (typeof ExpenseType)[keyof typeof ExpenseType]


export const ExpenseStatus = {
  paid: 'paid',
  pending: 'pending'
} as const

export type ExpenseStatus = (typeof ExpenseStatus)[keyof typeof ExpenseStatus]


export const IncomeType = {
  fixed: 'fixed',
  variable: 'variable'
} as const

export type IncomeType = (typeof IncomeType)[keyof typeof IncomeType]


export const InvoiceStatus = {
  open: 'open',
  closed: 'closed',
  overdue: 'overdue'
} as const

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus]


export const InvestmentType = {
  stock: 'stock',
  fixed_income: 'fixed_income',
  crypto: 'crypto',
  fund: 'fund'
} as const

export type InvestmentType = (typeof InvestmentType)[keyof typeof InvestmentType]